Verify flags is behind a sync.Once
Refactored common utilities to a new file
Removed separate verification for build flags for cleaner logic
