+++ kubetest2/kubetest2-gce/deployer/build.go
@@ -18,7 +18,6 @@ package deployer
 
 import (
 	"fmt"
-	"os"
 
 	"k8s.io/klog"
 	"k8s.io/test-infra/kubetest2/pkg/exec"
@@ -27,8 +26,8 @@ import (
 func (d *deployer) Build() error {
 	klog.Info("GCE deployer starting Build()")
 
-	if err := d.verifyBuildFlags(); err != nil {
-		return fmt.Errorf("Build() failed to verify build-specific flags: %s", err)
+	if err := d.verifyFlags(); err != nil {
+		return fmt.Errorf("Build() failed to verify flags: %s", err)
 	}
 
 	// this code path supports the kubernetes/cloud-provider-gcp build
@@ -46,24 +45,4 @@ func (d *deployer) Build() error {
 	// is set
 
 	return nil
-}
-
-func (d *deployer) setRepoPathIfNotSet() error {
-	if d.RepoRoot != "" {
-		return nil
-	}
-
-	path, err := os.Getwd()
-	if err != nil {
-		return fmt.Errorf("Failed to get current working directory for setting Kubernetes root path: %s", err)
-	}
-	klog.Infof("defaulting repo root to the current directory: %s", path)
-	d.RepoRoot = path
-
-	return nil
-}
-
-// verifyBuildFlags only checks flags that are needed for Build
-func (d *deployer) verifyBuildFlags() error {
-	return d.setRepoPathIfNotSet()
 }
