+++ kubetest2/kubetest2-gce/deployer/dumplogs.go
@@ -124,7 +124,6 @@ func (d *deployer) kubectlDump() error {
 	if err != nil {
 		return fmt.Errorf("couldn't use kubectl to dump cluster info: %s", err)
 	}
-	klog.Infof("About to run: %s", args)
 
 	return nil
 }
