+++ kubetest2/kubetest2-gce/deployer/up.go
@@ -66,52 +66,6 @@ func (d *deployer) Up() error {
 	return nil
 }
 
-func (d *deployer) buildEnv() []string {
-	// The base env currently does not inherit the current os env (except for PATH)
-	// because (for now) it doesn't have to. In future, this may have to change when
-	// support is added for k/k's kube-up.sh and kube-down.sh which support a wide
-	// variety of environment variables. Before doing so, it is worth investigating
-	// inheriting the os env vs. adding flags to this deployer on a case-by-case
-	// basis to support individual environment configurations.
-	var env []string
-
-	// path is necessary for scripts to find gsutil, gcloud, etc
-	// can be removed if env is inherited from the os
-	env = append(env, fmt.Sprintf("PATH=%s", os.Getenv("PATH")))
-
-	// used by config-test.sh to set $NETWORK in the default case
-	// if unset, bash's set -u gets angry and kills the log dump script
-	// can be removed if env is inherited from the os
-	env = append(env, fmt.Sprintf("USER=%s", os.Getenv("USER")))
-
-	// kube-up.sh, kube-down.sh etc. use PROJECT as a parameter for all gcloud commands
-	env = append(env, fmt.Sprintf("PROJECT=%s", d.GCPProject))
-
-	// kubeconfig is set to tell kube-up.sh where to generate the kubeconfig
-	// we don't want this to be the default because this kubeconfig "belongs" to
-	// the run of kubetest2 and so should be placed in the artifacts directory
-	env = append(env, fmt.Sprintf("KUBECONFIG=%s", d.kubeconfigPath))
-
-	// kube-up and kube-down get this as a default ("kubernetes") but log-dump
-	// does not. opted to set it manually here for maximum consistency
-	env = append(env, "KUBE_GCE_INSTANCE_PREFIX=kubetest2")
-	return env
-}
-
-func (d *deployer) verifyFlags() error {
-	if err := d.setRepoPathIfNotSet(); err != nil {
-		return err
-	}
-
-	d.kubectl = filepath.Join(d.RepoRoot, "cluster", "kubectl.sh")
-
-	if d.GCPProject == "" {
-		return fmt.Errorf("gcp project must be set")
-	}
-
-	return nil
-}
-
 func enableComputeAPI(project string) error {
 	// In freshly created GCP projects, the compute API is
 	// not enabled. We need it. Enabling it after it has
